requirements.txt
asgiref==3.6.0
Django==3.2.18
django-filter==23.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
psycopg2-binary==2.9.5
PyJWT==2.6.0
pytz==2023.3
sqlparse==0.4.3
